refactor(history): extract FacilityHistoryDao
