+++ application/src/main/java/fi/hsl/parkandride/core/back/FacilityRepository.java
@@ -5,14 +5,9 @@ package fi.hsl.parkandride.core.back;
 
 import fi.hsl.parkandride.core.domain.*;
 
-import java.util.List;
-
 public interface FacilityRepository {
 
     long insertFacility(Facility facility);
-
-    List<FacilityStatusHistory> getStatusHistory(long facilityId);
-    List<FacilityCapacityHistory> getCapacityHistory(long facilityId);
 
     void updateFacility(long facilityId, Facility facility);
